requirements.txt
oauthlib==3.2.2
openai==1.35.15
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0

